+++ hypixelio/__init__.py
@@ -1,22 +1,2 @@
-from .utils.helpers import (
-    form_url
-)
-
-from .models import (
-    boosters,
-    friends,
-    guild,
-    key,
-    player,
-    watchdog,
-)
-
-from .exceptions.exceptions import (
-    InvalidArgumentError,
-    HypixelAPIError,
-    RateLimitError,
-    PlayerNotFoundError,
-    GuildNotFoundError
-)
-
-from .client import Client
+from .lib import Client
+from .lib import Converters