+++ hypixelio/lib/__init__.py
@@ -1,0 +1,2 @@
+from .client import Client
+from .converters import Converters