+++ hypixelio/utils/constants.py
@@ -1,0 +1,10 @@
+"""
+This module is for defining the constant variables which will be
+used all over the code, hence reusing the same things, Instead of
+redefining, and accessing from a single place.
+"""
+
+HYPIXEL_API = "https://api.hypixel.net"
+MOJANG_API = "https://api.mojang.com"
+
+TIMEOUT = 10